bwz: Add a new tool
Bearbeitungswerkzeug is a tool for processing pcap files captured by
devlink-hwmsg.py. It's a shell filter that reads packets from one stream
and produces a stream of packets filtered and transformed from the
input. There's a simple DSL for selecting which packets to filter out
and which to keep, and a way to describe what information and in which
format should be included in the output stream.

Signed-off-by: Petr Machata <petrm@mellanox.com>
Signed-off-by: Jiri Pirko <jiri@mellanox.com>